services/requirements_stt.txt
amqp==5.3.1
billiard==4.2.1
blinker==1.9.0
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
kombu==5.4.2
MarkupSafe==3.0.2
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
six==1.17.0
tzdata==2025.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
prometheus-client==0.21.1
